refactor: rename TOC file to primarysidebar
